# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/gear/helmets": {
      "get": {
        "description": "Listado plano de HelmetModel con filtros administrativos avanzados (sin relaciones de variantes).",
        "operationId": "HelmetModelsAdminController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "brandSlug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "full_face",
                  "modular",
                  "open_face",
                  "half",
                  "off_road"
                ]
              }
            }
          },
          {
            "name": "shellMaterial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "polycarbonate",
                  "fiberglass",
                  "carbon_fiber",
                  "aramid_fiber"
                ]
              }
            }
          },
          {
            "name": "closureType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "double_d_ring",
                "micrometric",
                "quick_release",
                "magnetic",
                "pinch_buckle"
              ]
            }
          },
          {
            "name": "visorPinlockCompatible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "pinlock_30",
                  "pinlock_70",
                  "pinlock_70_max_vision",
                  "pinlock_120",
                  "pinlock_120_max_vision",
                  "pinlock_120_max_vision_xlt",
                  "pinlock_evo",
                  "pinlock_protectint",
                  "pinlock_max_vision",
                  "pinlock_ready",
                  "not_compatible"
                ]
              }
            }
          },
          {
            "name": "visorPinlockIncluded",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "tearOffCompatible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sunVisor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "intercomReady",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "removableLining",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "washableLining",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "emergencyRelease",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "maxSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "minWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "certification",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "dot",
                  "ece_2206",
                  "ece_2205",
                  "snell_m2020",
                  "snell_m2025",
                  "fim_frhphe_01",
                  "fim_frhphe_02"
                ]
              }
            }
          },
          {
            "name": "minCreatedAt",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxCreatedAt",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "minUpdatedAt",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxUpdatedAt",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "onlyDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingSunVisorType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingIntercomDesignedBrand",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingIntercomDesignedModel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "missingPinlockDksCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de modelos (vista admin, sin variantes)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Listar modelos de casco",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "post": {
        "operationId": "HelmetModelsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Casco creado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Crear modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/admin/gear/helmets/{id}": {
      "get": {
        "description": "Incluye todos los campos internos y variantes eliminadas.",
        "operationId": "HelmetModelsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle completo del casco (vista admin)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener casco por ID",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "put": {
        "description": "Todos los campos son opcionales (actualización parcial).",
        "operationId": "HelmetModelsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Casco actualizado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "delete": {
        "description": "Marca el casco como eliminado sin borrarlo de la base de datos. Restaurable con /restore.",
        "operationId": "HelmetModelsAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Casco eliminado (soft delete)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar casco (soft delete)",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/admin/gear/helmets/{id}/restore": {
      "post": {
        "description": "Revierte el soft delete, el casco vuelve al estado activo.",
        "operationId": "HelmetModelsAdminController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Casco restaurado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Restaurar casco eliminado",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moto Gear Picker API",
    "description": "API para selección y gestión de equipamiento de moto",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "full_face",
                "modular",
                "open_face",
                "half",
                "off_road"
              ]
            }
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "polycarbonate",
                "fiberglass",
                "carbon_fiber",
                "aramid_fiber"
              ]
            }
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlockCompatible": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "pinlock_30",
                "pinlock_70",
                "pinlock_70_max_vision",
                "pinlock_120",
                "pinlock_120_max_vision",
                "pinlock_120_max_vision_xlt",
                "pinlock_evo",
                "pinlock_protectint",
                "pinlock_max_vision",
                "pinlock_ready",
                "not_compatible"
              ]
            }
          },
          "visorPinlockIncluded": {
            "type": "boolean"
          },
          "pinlockDksCode": {
            "type": "string"
          },
          "tearOffCompatible": {
            "type": "boolean"
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "ece_2206",
                "ece_2205",
                "snell_m2020",
                "snell_m2025",
                "fim_frhphe_01",
                "fim_frhphe_02"
              ]
            }
          },
          "includedAccessories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "slug",
          "brandId",
          "helmetType"
        ]
      },
      "UpdateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "full_face",
                "modular",
                "open_face",
                "half",
                "off_road"
              ]
            }
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "polycarbonate",
                "fiberglass",
                "carbon_fiber",
                "aramid_fiber"
              ]
            }
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlockCompatible": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "pinlock_30",
                "pinlock_70",
                "pinlock_70_max_vision",
                "pinlock_120",
                "pinlock_120_max_vision",
                "pinlock_120_max_vision_xlt",
                "pinlock_evo",
                "pinlock_protectint",
                "pinlock_max_vision",
                "pinlock_ready",
                "not_compatible"
              ]
            }
          },
          "visorPinlockIncluded": {
            "type": "boolean"
          },
          "pinlockDksCode": {
            "type": "string"
          },
          "tearOffCompatible": {
            "type": "boolean"
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "ece_2206",
                "ece_2205",
                "snell_m2020",
                "snell_m2025",
                "fim_frhphe_01",
                "fim_frhphe_02"
              ]
            }
          },
          "includedAccessories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}